refactor(daemon): extract `Daemon` into separate module
